+++ tests/core/resource/test_views.py
@@ -3,7 +3,8 @@ import uuid
 
 from kinto.core.storage import exceptions as storage_exceptions
 from kinto.core.errors import ERRORS
-from kinto.core.testing import unittest
+from kinto.core.testing import unittest, FormattedErrorMixin
+
 
 from ..support import BaseWebTest
 
@@ -776,3 +777,14 @@ class SchemaLessPartialResponseTest(BaseWebTest, unittest.TestCase):
         resp = self.app.get(self.collection_url + '?_fields=nationality')
         result = resp.json['data'][0]
         self.assertNotIn('nationality', result)
+
+
+class UnicodeDecodeErrorTest(BaseWebTest, FormattedErrorMixin, unittest.TestCase):
+    collection_url = '/spores'
+
+    def test_wrong_filter_encoding_raise_a_400_bad_request(self):
+        resp = self.app.get(self.collection_url + '?foo\xe2\xfc\xa7bar',
+                            headers=self.headers, status=400)
+        self.assertFormattedError(resp, 400, ERRORS.INVALID_PARAMETERS, "Bad Request",
+                                  "A request with an incorrect encoding in the querystring was"
+                                  "received. Please make sure your requests are encoded in UTF-8")