+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ This document describes changes between each past release.
 **Bug fixes**
 
 - Return a 400 Bad Request instead of crashing when the querystring contains bad characters. (fixes #1195)
+- Fixed Postgresql errors when specifying empty values in querystring numeric filters. (fixes #1194)
+
 
 6.0.5 (2017-04-12)
 ------------------